requirements.txt
# Annotation Checker 依赖

# HTTP 请求
requests>=2.28.0

# YAML 配置解析
pyyaml>=6.0

# SSH/SFTP 客户端
paramiko>=3.0.0

# 数值计算
numpy>=1.21.0
